migrated admin post requests to postrequests.ts
